+++ CHANGELOG.md
@@ -8,10 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2023-12-29
 ### Added
 - Added `Button.tsx` for a CustomButton 
+- Added width in `Button.tsx` to handle small screens
 
 ### Changed
 - Changed the Colorpalette of Buttons and Navmenu on mobile and big screens
 - Changed the size of Titles to a smaller size
+- Changed Buttons of `VillaAnnaCountdownPage` with new CustomButtons
+
+### Fixed
+- When in sublink of /villa-anna/ Route, the button remains active
 
 ## [v15.9.0] - 2023-12-28
 ### Added